Crate swarm

An object pooling system optimized for perfomance.

The pooling system manages object instances of a cutom type, and provides update loops to over them.

Examples

extern crate swarm;
use swarm::Swarm;
use swarm::tools::byte_str::ByteStr;
 
// create an object you want to pool
#[derive(Default, Copy, Clone)]     // The object you want to pool must implement these
pub struct MyPoolObject {           // Swarm uses Copy and therfore only accepts Sized properties!
    pub name: ByteStr,              // This means types such as String and Vec aren't allowed
    pub value: usize,               // The tools module has a few tools that deal with this
}
 
// create properties you want to share with pooled objects
pub struct MySwarmProperties;
 
fn main() {
    let swarm = Swarm::<MyPoolObject, MySwarmProperties>::new(10, MySwarmProperties);
    assert!(swarm.capacity() == 10);
}

Modules

control
Controling a swarm during the swarm.update() loop

tools
Tools contains a set of Sized types that can be useful when creating pool objects. Beacuse pool objects must Sized, all of their properties must be Sized.

types
Types used by the Swarm pool.

Structs

Swarm
The actual Swarm pool