Struct swarm::Swarm

pub struct Swarm<ItemType, Properties> {
    pub properties: Properties,
    // some fields omitted
}

The actual Swarm pool

Fields
properties: Properties
Implementations

impl<ItemType: Default + Copy, Properties> Swarm<ItemType, Properties>

pub fn new(capacity: usize, properties: Properties) -> Self

Create a new Swarm object pool

The maximum capacity of the pool is defined here and cannot be changed aferwards, This is the number of instances you can span.

You can add custom swarm properties for sharing values between spawned instances while iterating over these spawns.

Example

extern crate swarm;
use swarm::Swarm;
use swarm::tools::byte_str::ByteStr;
 
// create an object you want to pool
// Swarm uses Copy and therfore only accepts Sized properties!
// This means types such as String and Vec aren't allowed
// The tools module has a few tools that deal with this
#[derive(Default, Copy, Clone)]     
pub struct MyPoolObject {           
    pub name: ByteStr,              
    pub value: usize,               
}
 
// create properties you want to share with pooled objects
pub struct MySwarmProperties;
 
fn main() {
    let swarm = Swarm::<MyPoolObject, MySwarmProperties>::new(10, MySwarmProperties);
    assert!(swarm.capacity() == 10);
}

pub fn spawn(&mut self) -> Option<Spawn>

Create a new pool instance. Spawns are included in the update loop provided by Swarm

Example

extern crate swarm;
use swarm::Swarm;
use swarm::tools::byte_str::ByteStr;
 
// create an object you want to pool
#[derive(Default, Copy, Clone)] 
pub struct MyPoolObject {
    pub name: ByteStr,
    pub value: usize,
}
 
// create properties you want to share with pooled objects
pub struct MySwarmProperties;
 
fn main() {
    let mut swarm = Swarm::<MyPoolObject, MySwarmProperties>::new(10, MySwarmProperties);
    let spawn = swarm.spawn();
 
    assert!(spawn.is_some());
    assert_eq!(swarm.count(), 1);
}

pub fn kill(&mut self, target: &Spawn)

Remove a spawn from the swarm pool

Example

extern crate swarm;
use swarm::Swarm;
use swarm::tools::byte_str::ByteStr;
 
// create an object you want to pool
#[derive(Default, Copy, Clone)] 
pub struct MyPoolObject {
    pub name: ByteStr,
    pub value: usize,
}
 
// create properties you want to share with pooled objects
pub struct MySwarmProperties;
 
fn main() {
    let mut swarm = Swarm::<MyPoolObject, MySwarmProperties>::new(10, MySwarmProperties);
    let spawn = swarm.spawn();
 
    assert!(spawn.is_some());
    assert_eq!(swarm.count(), 1);
 
    swarm.kill(&spawn.unwrap());
    assert_eq!(swarm.count(), 0);
}

pub fn find_spawn(&self, pos: &ObjectPosition) -> Spawn

Returns a spawn reference object from an object position within the pool

pub fn fetch(&mut self, spawn: &Spawn) -> &mut ItemType

Returns a mutable reference to an object from the Swarm pool. The supplied Spawn reference object points out which object to return.

pub fn fetch_ref(&self, spawn: &Spawn) -> &ItemType

Returns a immutable reference to an object from the Swarm pool. The supplied Spawn reference object points out which object to return.

pub fn fetch_raw(&mut self, pos: &ObjectPosition) -> &mut ItemType

Returns a mutable reference to an object from the Swarm pool. The supplied object position points out which object to return.

NOTE: when killing objects the order of object positions change, Use fetch() instead of fetch_raw() if you are going to kill spawns at any point in your code.

pub fn count(&self) -> usize

Returns the number of spawned instances currently availeble

pub fn capacity(&self) -> usize

Returns the maximum number of instances that can be spawned

pub fn for_each(&mut self, handler: ForEachHandler<ItemType>)

Loops through all spawned instances and returns them via a callback handler. The callback handler is supplied with a mutable reference of these instances so that the object data of each looped instance can be changed.

pub fn for_all(&mut self, handler: ForAllHandler<ItemType, Properties>)

Loops through all spawned instances and returns their object position via a callback handler. The callback handler also hands out a mutable reference to the object pool and the swarm properties object.

NOTE: This methode is quite fast (faster than the foreach or update methodes), but at the cost that is uses object positions. When an object is killed, the order of some object positions are changed. Because of that, no tools for spawning nor killing are supplied, which makes it impossible to do so within this loop.

pub fn update(&mut self, handler: UpdateHandler<ItemType, Properties>)

Loops through all spawned instances and returns a SwarmControl object via a callback handler. The swarm control objects lets you edit the currently updated object as well as spawning and killing instances.

NOTE: This methode is slower then the other loops, but gives you full control over the swarm.