Struct swarm::swarm::Swarm

pub struct Swarm {
    config: Config,
    members: Vec<Member>,
}

Keeps track of necessary parameters for the load testing

Fields

config

Config

members

List of members

Methods

impl Swarm

fn new(config: Config) -> Swarm

Instantiates a new swarm with the given settings

fn unleash(&mut self) -> Result<(), ()>

Generates all the necessary threads and sends the requests

fn stats(&self) -> Stats

Returns the swarm stats