Trait swarm::tools::sized_pool::StackPool

pub trait StackPool<ItemType> {
    fn count(&self) -> &usize;

    fn count_mut(&mut self) -> &mut usize;

    fn ref_at(&self, item_index: &usize) -> &Option<ItemType>;

    fn ref_sorted(&self, ord_index: &usize) -> &Option<ItemType>;

    fn item_at(&mut self, item_index: &usize) -> &mut Option<ItemType>;

    fn item_last(&mut self) -> &mut Option<ItemType>;

    fn item_sorted(&mut self, ord_index: &usize) -> &mut Option<ItemType>;

    fn order_at(&mut self, ord_index: &usize) -> &mut usize;
}

The StackPool trait is used by this module to communicate with the different pool types on a polymorphic level.

It was not meant to be public, but hiding the ugly beast would dissapoint the compiler when used externally.

Required methods

fn count(&self) -> &usize

fn count_mut(&mut self) -> &mut usize

fn ref_at(&self, item_index: &usize) -> &Option<ItemType>

fn ref_sorted(&self, ord_index: &usize) -> &Option<ItemType>

fn item_at(&mut self, item_index: &usize) -> &mut Option<ItemType>

fn item_last(&mut self) -> &mut Option<ItemType>

fn item_sorted(&mut self, ord_index: &usize) -> &mut Option<ItemType>

fn order_at(&mut self, ord_index: &usize) -> &mut usize

Loading content...